requirements.txt
fastapi==0.115.8
pydantic==2.10.6
Pillow==11.1.0
numpy==2.2.2
scikit-image==0.25.1
uvicorn==0.34.0
